Extracted reusable methods into mixin
